Turned three examples into tests
chapel.sfile-inline.py is now test_unusual_sfile_loc.py, chapel.inline.begin.py
is now test_begin.py, and chapel.inline.cobegin.py is now test_cobegin.py
